% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webdeveloper.R
\name{endServer}
\alias{endServer}
\title{Stop HTTP server(s) by calling httpuv::stopServer() or httpuv::stopAllServers().}
\usage{
endServer(x = NULL, all = FALSE)
}
\arguments{
\item{x}{A server object that was previously returned from serveHTTP.}

\item{all}{TRUE/FALSE, if TRUE, calls httpuv::stopAllServers.}
}
\value{
Nothing.
}
\description{
Stop HTTP server(s) by calling httpuv::stopServer() or httpuv::stopAllServers().
}
\examples{
endServer(all = TRUE)
}
