% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webdeveloper.R
\name{findTemplateVars}
\alias{findTemplateVars}
\title{Find the names of any placeholder variables that exist in a HTML document string.}
\usage{
findTemplateVars(x)
}
\arguments{
\item{x}{HTML string to check for placeholder.}
}
\value{
A vector of the names of template vars found in the HTML string.
}
\description{
Find the names of any placeholder variables that exist in a HTML document string.
}
\examples{
findTemplateVars(x = html(body(templateVar("body_var"))))
}
