% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webdeveloper.R
\name{fromInput}
\alias{fromInput}
\title{Prepare values collected from HTML forms to save to a SQL database by calling quoteText. If x is "", returns "NULL".}
\usage{
fromInput(x)
}
\arguments{
\item{x}{A vector of length 1.}
}
\value{
A string, if x is "", returns "NULL".
}
\description{
Prepare values collected from HTML forms to save to a SQL database by calling quoteText. If x is "", returns "NULL".
}
\examples{
fromInput("Test")
fromInput("100")
fromInput(100)
fromInput("")
}
