% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webdeveloper.R
\name{idAddAffixes}
\alias{idAddAffixes}
\title{Add a prefix and suffix to an id}
\usage{
idAddAffixes(prefix, id, suffix, prefix_sep = "X", suffix_sep = "-")
}
\arguments{
\item{prefix}{A string, the prefix to add.}

\item{id}{A string to add a prefix and suffix to.}

\item{suffix}{A string, the suffix to add.}

\item{prefix_sep}{A string, the prefix separator to use.
This should be different than suffix_sep.}

\item{suffix_sep}{A string, the suffix separator to use.
This should be different than prefix_sep.}
}
\value{
A string.
}
\description{
Add a prefix and suffix to an id
}
\examples{
idAddAffixes("group1", "example", 1)
}
