% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webdeveloper.R
\name{idAddPrefix}
\alias{idAddPrefix}
\title{Add a prefix to an id}
\usage{
idAddPrefix(prefix, id, sep = "X")
}
\arguments{
\item{prefix}{A string, the prefix to add.}

\item{id}{A string to add a prefix to.}

\item{sep}{A string, the separator to use.}
}
\value{
A string.
}
\description{
Add a prefix to an id
}
\examples{
idAddSuffix("example", 1)
}
