% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webdeveloper.R
\name{idAddSuffix}
\alias{idAddSuffix}
\title{Add a suffix to an id}
\usage{
idAddSuffix(id, suffix, sep = "-")
}
\arguments{
\item{id}{A string to add a suffix to.}

\item{suffix}{A string, the suffix to add.}

\item{sep}{A string, the separator to use.}
}
\value{
A string.
}
\description{
Add a suffix to an id
}
\examples{
idAddSuffix("example", 1)
}
