% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webdeveloper.R
\name{idParseAffixes}
\alias{idParseAffixes}
\title{Remove a prefix and suffix from an id}
\usage{
idParseAffixes(id, split = "X|-")
}
\arguments{
\item{id}{A string to remove a prefix and suffix from.}

\item{split}{A regular expression to use for splitting the prefix and suffix from the id.}
}
\value{
A named vector, with prefix, id, and suffix returned in that order.
}
\description{
Remove a prefix and suffix from an id
}
\examples{
idParseAffixes(idAddAffixes("group1", "example", 1))
}
