% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webdeveloper.R
\name{idParsePrefix}
\alias{idParsePrefix}
\title{Remove a prefix from an id}
\usage{
idParsePrefix(id, split = "X", position = 2)
}
\arguments{
\item{id}{A string to remove a prefix from.}

\item{split}{A string, the separator to use for splitting the id.}

\item{position}{A integer vector, the position of the split string to return.}
}
\value{
A vector.
}
\description{
Remove a prefix from an id
}
\examples{
idParsePrefix(idAddPrefix("example", 1))
}
