% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webdeveloper.R
\name{idParseSuffix}
\alias{idParseSuffix}
\title{Remove a suffix from an id}
\usage{
idParseSuffix(id, split = "-", position = 1)
}
\arguments{
\item{id}{A string to remove a suffix from.}

\item{split}{A string, the separator to use for splitting the id.}

\item{position}{A integer vector, the position of the split string to return.}
}
\value{
A vector.
}
\description{
Remove a suffix from an id
}
\examples{
idParseSuffix(idAddSuffix("example", 1))
}
