% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webdeveloper.R
\name{parseContentTypeHeader}
\alias{parseContentTypeHeader}
\title{Parse the content type header string to return the content type and boundary}
\usage{
parseContentTypeHeader(x)
}
\arguments{
\item{x}{A string containing the content type header.}
}
\value{
A named list with "content_type" and "boundary" if boundary is present.
}
\description{
Parse the content type header string to return the content type and boundary
}
\examples{
parseContentTypeHeader("application/x-www-form-urlencoded")
}
