% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webdeveloper.R
\name{parseHTTP}
\alias{parseHTTP}
\title{Parse a HTTP request}
\usage{
parseHTTP(x, content_type_header = NULL, consolidate = TRUE)
}
\arguments{
\item{x}{The body of the HTTP request}

\item{content_type_header}{A string containing the content type header.}

\item{consolidate}{TRUE/FALSE, if TRUE, consolidates items with the same name.}
}
\value{
A named list.
}
\description{
Parse a HTTP request
}
\examples{
parseHTTP("?form_id=example&col_name=Test+String", "application/x-www-form-urlencoded")
}
