% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webdeveloper.R
\name{parseMultiPartFormParams}
\alias{parseMultiPartFormParams}
\title{Helper function for parseMultiPartFormData}
\usage{
parseMultiPartFormParams(x)
}
\arguments{
\item{x}{A vector, a chunk of multi-part form data to parse.}
}
\value{
A named list.
}
\description{
Helper function for parseMultiPartFormData
}
\examples{
parseMultiPartFormParams(c("Content-Disposition: form-data; name=\"form_name\"", "", "Example"))
}
