% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webdeveloper.R
\name{parseQueryString}
\alias{parseQueryString}
\title{Parse a query string}
\usage{
parseQueryString(x, split = "&", consolidate = TRUE)
}
\arguments{
\item{x}{A string containing the query string.}

\item{split}{A string, the character to split by.}

\item{consolidate}{TRUE/FALSE, if TRUE, consolidates items with the same name.}
}
\value{
A named list.
}
\description{
Parse a query string
}
\examples{
parseQueryString("?form_id=example&col_name=Test+String")
}
