% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webdeveloper.R
\name{sampleStr}
\alias{sampleStr}
\title{Generates (pseudo)random strings of the specified char length.}
\usage{
sampleStr(char)
}
\arguments{
\item{char}{A integer, the number of chars to include in the output string.}
}
\value{
A string.
}
\description{
Generates (pseudo)random strings of the specified char length.
}
\examples{
sampleStr(10)
}
