% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webdeveloper.R
\name{templateVar}
\alias{templateVar}
\title{Create a string to use as a placeholder variable in a HTML document.}
\usage{
templateVar(x)
}
\arguments{
\item{x}{Name of placeholder.}
}
\value{
A string.
}
\description{
Create a string to use as a placeholder variable in a HTML document.
}
\examples{
templateVar("my_dynamic_var")
}
