% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webdeveloper.R
\name{toInput}
\alias{toInput}
\title{Replace NA values with "", useful for passing values to HTML tags.}
\usage{
toInput(x)
}
\arguments{
\item{x}{A vector of length 1.}
}
\value{
A string, if x is NA, returns "".
}
\description{
Replace NA values with "", useful for passing values to HTML tags.
}
\examples{
toInput(NA)
}
